type double

+++ src/GraphAlgorithms/DistanceHeuristic/DistanceHeuristic.java
@@ -1,6 +1,6 @@
 package GraphAlgorithms.DistanceHeuristic;
 
 public interface DistanceHeuristic<Stop> {
-    public int distance(Stop v0, Stop v1);
+    public double distance(Stop v0, Stop v1);
 }
